refactor: added debug logging to services
This is to help provide which services were added to the engine, and which ones were able to be initialized from the container before the engine tries to start the game.

Side bonus, this also pre-loads all static services before the game starts, reducing lag if they are called mid-game loop. And yes I know that also means they will always be loaded even if not needed.